doc: example of HHN with network permutations
The file "./examples/example_commands_permhierarchies.sh" contains
a modified version of "./examples/example_commands.sh" using the
permutation of the network's edges for the test (instead of
permuting scores).

It runs completely (currently only using 25 permutations) only
printing a warning for a deprecation of a numpy method that is
used in "./src/hhio.py".